refactor: 의미를 명확하게하기 위해 패키지 추가 및 Application의 이름 수정
