+++ src/main/java/Janggi/piece/Country.java
@@ -1,0 +1,16 @@
+package Janggi.piece;
+
+public enum Country {
+    HAN, CHO;
+
+    public Country toggleCountry() {
+        if (this == HAN) {
+            return CHO;
+        }
+        return HAN;
+    }
+
+    public static Country getFirstTurnCountry(){
+        return CHO;
+    }
+}
