+++ src/main/java/Janggi/piece/Horse.java
@@ -1,0 +1,28 @@
+package Janggi.piece;
+
+import Janggi.board.Board;
+import Janggi.board.Position;
+
+public class Horse extends Piece {
+
+    public Horse(final Country country) {
+        super(country);
+    }
+
+    @Override
+    public boolean canMove(final Position now, final Position destination, final Board board) {
+        if (now.calculateDistance(destination) != Math.sqrt(5)) {
+            return false;
+        }
+        final Position position = now.calculateHorseMiddlePosition(destination);
+        if (board.existPieceByPosition(position)) {
+            return false;
+        }
+        return true;
+    }
+
+    @Override
+    public boolean equalsType(final Piece piece) {
+        return piece instanceof Horse;
+    }
+}
