--- src/main/java/board/Board.java
@@ -1,65 +1,0 @@
-package board;
-
-import java.util.HashMap;
-import java.util.Map;
-import piece.Piece;
-import piece.Country;
-
-public class Board {
-
-    private final Map<Position, Piece> janggiBoard;
-
-    public Board(final Map<Position, Piece> janggiBoard) {
-        this.janggiBoard = new HashMap<>(janggiBoard);
-    }
-
-    public boolean existPieceByPosition(final Position existPosition) {
-        return janggiBoard.containsKey(existPosition);
-    }
-
-    public boolean equalsTypeByPositionAndPiece(final Position position, final Piece piece) {
-        if (janggiBoard.containsKey(position)) {
-            final Piece target = janggiBoard.get(position);
-            return target.equalsType(piece);
-        }
-        return false;
-    }
-
-    public boolean equalsTeamTypeByPosition(final Position position, final Country country) {
-        if (janggiBoard.containsKey(position)) {
-            final Piece piece = janggiBoard.get(position);
-            return piece.equalsTeamType(country);
-        }
-        return false;
-    }
-
-    public void updatePosition(final Position source, final Position destination, final Country country) {
-        validatePositionAndTeam(source, country);
-        validatePieceCanMove(source, destination);
-
-        movePieceToDestination(source, destination);
-    }
-
-    private void validatePositionAndTeam(final Position source, final Country country) {
-        if (!janggiBoard.containsKey(source) || !janggiBoard.get(source).equalsTeamType(country)) {
-            throw new IllegalArgumentException("scr 좌표에 기물이 존재하지 않거나, 해당 팀의 기물이 아닙니다.");
-        }
-    }
-
-    private void validatePieceCanMove(final Position source, final Position destination) {
-        final Piece piece = janggiBoard.get(source);
-
-        if (!piece.isAbleToMove(source, destination, this)) {
-            throw new IllegalArgumentException("해당 기물은 해당 위치로 이동할 수 없습니다.");
-        }
-    }
-
-    private void movePieceToDestination(final Position source, final Position destination) {
-        janggiBoard.put(destination, janggiBoard.get(source));
-        janggiBoard.remove(source);
-    }
-
-    public Map<Position, Piece> getJanggiBoard() {
-        return new HashMap<>(janggiBoard);
-    }
-}
