--- src/main/java/piece/Country.java
@@ -1,16 +1,0 @@
-package piece;
-
-public enum Country {
-    HAN, CHO;
-
-    public Country toggleCountry() {
-        if (this == HAN) {
-            return CHO;
-        }
-        return HAN;
-    }
-
-    public static Country getFirstTurnCountry(){
-        return CHO;
-    }
-}
