--- src/main/java/piece/Guard.java
@@ -1,21 +1,0 @@
-package piece;
-
-import board.Board;
-import board.Position;
-
-public class Guard extends Piece{
-
-    public Guard(final Country country) {
-        super(country);
-    }
-
-    @Override
-    public boolean canMove(final Position now, final Position destination, final Board board) {
-        return now.calculateDistance(destination) == 1;
-    }
-
-    @Override
-    public boolean equalsType(final Piece piece) {
-        return piece instanceof Guard;
-    }
-}
