--- src/test/java/piece/GuardTest.java
@@ -1,56 +1,0 @@
-package piece;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import board.Board;
-import board.Position;
-import java.util.HashMap;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class GuardTest {
-
-    @DisplayName("General은 주변 한칸으로 이동할 수 있다.")
-    @Test
-    void general() {
-        // given
-        final Piece generalPiece = new General(Country.CHO);
-        final Position now = new Position(1, 1);
-        final Position ableDest = new Position(1, 2);
-        final Position notAbleDest = new Position(1, 3);
-        final Board board = new Board(new HashMap<>());
-        final Country country = Country.HAN;
-
-        // when
-        final boolean actual1 = generalPiece.isAbleToMove(now, ableDest, board);
-        final boolean actual2 = generalPiece.isAbleToMove(now, notAbleDest, board);
-
-        // then
-        org.junit.jupiter.api.Assertions.assertAll(
-                () -> assertThat(actual1).isTrue(),
-                () -> assertThat(actual2).isFalse()
-        );
-    }
-
-    @DisplayName("guard는 주변 한칸으로 이동할 수 있다.")
-    @Test
-    void guard() {
-        // given
-        final Piece guardPiece = new Guard(Country.CHO);
-        final Position now = new Position(1, 1);
-        final Position ableDest = new Position(1, 2);
-        final Position notAbleDest = new Position(1, 3);
-        final Board board = new Board(new HashMap<>());
-        final Country country = Country.HAN;
-
-        // when
-        final boolean actual1 = guardPiece.isAbleToMove(now, ableDest, board);
-        final boolean actual2 = guardPiece.isAbleToMove(now, notAbleDest, board);
-
-        // then
-        org.junit.jupiter.api.Assertions.assertAll(
-                () -> assertThat(actual1).isTrue(),
-                () -> assertThat(actual2).isFalse()
-        );
-    }
-}
